Fix pdf file upload
- Displays original filename when uploading pdf file
