+++ server/controllers/pdfController.js
@@ -67,32 +67,18 @@ const getAllPdf = (req, res) => {
     if (!user) {
       return res.sendStatus(401).send('The user does not exist.');
     }
-    Pdf.find({user: user._id})
-      .distinct('fileId')
-      .then(function (pdf) {
-        console.log(pdf);
-        gfs.files.find({_id: {$in: pdf}}).toArray((err, files) => {
-          if (!files || files.length === 0) {
-            return res.json({
-              files: false,
-            });
-          }
-          files.map((file) => {
-            if (file.contentType === 'application/pdf') {
-              file.isPDF = true;
-            } else {
-              file.isPDF = false;
-            }
-          });
-          const pdfObj = [];
-          for (file of files) {
-            if (file.isPDF) {
-              pdfObj.push(file);
-            }
-          }
-          return res.json({files: pdfObj});
+    const result = [];
+
+    Pdf.find({user: user._id}).then((pdfs) => {
+      for (ele of pdfs) {
+        result.push({
+          originalname: ele.originalName,
+          getFileLink: '/api/pdf/' + ele.filename,
+          deleteFileLink: '/pdf/' + ele.fileId,
         });
-      });
+      }
+      return res.json({pdfs: result});
+    });
   });
 };
