+++ server/models/pdfModel.js
@@ -11,6 +11,8 @@ const pdfSchema = new Schema(
       type: mongoose.Schema.Types.ObjectId,
       ref: 'User',
     },
+    originalName: String,
+    filename: String,
   },
   {timestamps: true}
 );
